+++ utils/mini_map.c
@@ -6,7 +6,7 @@
 /*   By: tjamis <tjamis@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/11 16:30:11 by tjamis            #+#    #+#             */
-/*   Updated: 2022/04/11 18:59:07 by tjamis           ###   ########.fr       */
+/*   Updated: 2022/04/11 20:17:40 by tjamis           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ int	max_len(char **map)
 	max = 0;
 	while (map && map[i])
 	{
-		if (max < ft_strlen(map[i]))
+		if (max < (int)ft_strlen(map[i]))
 			max = ft_strlen(map[i]);
 		i++;
 	}
